+++ addon/components/modal-dialog/index.js
@@ -2,7 +2,7 @@ import Component from '@glimmer/component';
 import ModalDialogHeader from './header';
 import ModalDialogContent from './content';
 import ModalDialogFooter from './footer';
-import { Promise, resolve } from 'rsvp';
+import { resolve, defer } from 'rsvp';
 import { action } from '@ember/object';
 import { tracked } from '@glimmer/tracking';
 import { scheduleOnce } from '@ember/runloop';
@@ -10,6 +10,7 @@ import { disableBodyScroll, enableBodyScroll } from 'body-scroll-lock';
 
 export default class ModalDialogComponent extends Component {
   element = null;
+  willAnimate = defer();
 
   ModalDialogHeader = ModalDialogHeader;
   ModalDialogContent = ModalDialogContent;
@@ -70,11 +71,22 @@ export default class ModalDialogComponent extends Component {
     });
   }
 
+  _hide() {
+    this.isShowing = false;
+
+    return this._waitForAnimation();
+  }
+
   @action
   warn() {
     this.isWarning = true;
 
     this._waitForAnimation().then(() => (this.isWarning = false));
+  }
+
+  @action
+  handleAnimationEnd() {
+    this.willAnimate.resolve();
   }
 
   @action
@@ -125,12 +137,6 @@ export default class ModalDialogComponent extends Component {
       .then((data) => this.args.onLoaded?.(data))
       .catch((error) => this.args.onLoadError?.(error))
       .finally(() => (this.isLoading = false));
-  }
-
-  _hide() {
-    this.isShowing = false;
-
-    return this._waitForAnimation();
   }
 
   _watchForContentChanges() {
@@ -193,11 +199,8 @@ export default class ModalDialogComponent extends Component {
   }
 
   _waitForAnimation() {
-    return new Promise((resolve) => {
-      this.element.addEventListener('animationend', resolve, {
-        once: true
-      });
-    });
+    this.willAnimate = defer();
+    return this.willAnimate.promise;
   }
 
   _pressedEscape(e) {